don't show icons on the desktop by default
The plan for GNOME 3 sessions is to use nautilus in an application mode
for file manipulation.  We are moving away from file management being
primary and core functionality of the desktop shell.  And consequently
nautilus is no longer considered to be part of the GNOME 3 shell - either
in default or fallback modes.