schemas: Add new 'switch-applications' keybinding
GNOME Shell currently uses the 'switch-windows' keybinding for its
application switcher, although it does not switch between windows
but applications (duh!).
Add a new dedicated keybinding instead, which will free 'switch-windows'
to be used for window switching again. People unhappy with the Shell's
alt-tab popup may then use it instead of (or in addition to!) the
application-based switcher.

https://bugzilla.gnome.org/show_bug.cgi?id=688913