Extend the app-folders schema
To replace the previous functionality of having default folders
for (some) categories, we need add some more keys to app-folder
schema.

The categories key defines a list of categories that should be
put into the folder by default.

The excluded-apps key provides a way to override the placement
by categories for individual apps.

https://bugzilla.gnome.org/show_bug.cgi?id=722658