schemas: Add a way to turn off "PC-style" right-click
On Macs, there's no physical touchpad buttons, and right clicks are
performed through 2-finger clicks.

On Lenovos (and many other PCs), specific areas of the device are set
aside to act as buttons.

The default configuration for those is to leave it as is the default for
that specific hardware, but allows to remove button areas on PC laptops.

See equivalent, old-style work-around:
https://gist.github.com/hadess/6779414