MSVC builds: Add common autotools module for introspection
This adds a common autotools module that can be included in the
Makefile.am's to generate the file list for files to introspect,
and the g-ir-scannner command to generate the .gir files, and to
compile the corresponding .typelib files.

This is written with reusability in mind, so this module is copied from
gobject-introspection's $(srcroot)/build.

https://bugzilla.gnome.org/show_bug.cgi?id=723734