schemas: Change default click method for touchpads
Use "Mac-style" click method for touchpads, with the whole touchpad
being available for navigating, two and 3-finger clicks are used to
emulate the right and middle buttons.

The default setting on "Windows" machines is a remnant of the olden
crappy touchpads days where the hardware vendor would have removed the
hardware buttons but didn't have multi-touch support in the touchpad,
likely in addition to Windows not supporting multi-touch gestures well
enough for those devices.

This does not change the behaviour of physical buttons associated with
touchpads or trackpoints, or the click regions on touchpads for
touchpads not capable of multi-touch clicks.

https://bugzilla.gnome.org/show_bug.cgi?id=757255