schemas: Scroll method key changes again
Some users, even with touchpads capable of two-finger scrolling do not
like two-finger scrolling, and prefer edge scrolling, whether simply
because it "does not feel right" or because it can cause physical pain.

So instead of forcing two-finger scrolling when the touchpad is capable,
enable it as a separate setting, and leave the unfancied but required
edge-scrolling disabled by default. This will require users to enable
edge-scrolling by hand, not too dissimilarly to other universal access
features.

https://bugzilla.gnome.org/show_bug.cgi?id=768244