Revert "Add an alternative shortcut for 'go to overview'"
This reverts commit 8808a8fdba48fb55de84de3e98e081b38f2fd92d.

The same shortcut - <super>s - is used for two different actions, which
doesn't work - after registering the first shortcut, grabbing the second
one will fail.
That is, only one of the 'panel-main-menu' and 'toggle-overview' action
will actually work. This doesn't matter when the apps-menu extension is
not used, because the actions do the same thing in that case, but in
the case where the newly added action is actually useful, <super>s will
only work as expected when 'toggle-overview' is the "winning" action.

The fix here is to revert commit 8808a8fdba48, so we have <alt>f1 for
'panel-main-menu' and <super>s for 'toggle-overview' - that way,
<super>s will always toggle the overview, and <alt>f1 will either
act as an alternative or toggle the apps menu.

With help from Florian Müllner <florian.muellner@gmail.com>

Conflicts:
	schemas/org.gnome.desktop.wm.keybindings.gschema.xml.in

https://bugzilla.gnome.org/show_bug.cgi?id=720994