MSVC builds: Add --datadir to genschema.py
It may be the case that the user wants to use a shared location for the
datadir that is not bound to a particular prefix, so we provide a
--datadir option for one to specify such a path, which will override the
default $(PREFIX)\share in the generated schema files.

Also deduce the URI-style of local on-disk paths directly in this
script when we generate the gschema files, so that we can remove
uri.py as a result, since it didn't have to be a file on its own
after all.