build: Honor DATADIR for Visual Studio builds
With this, we pass --datadir=$(DATADIR) to genschema.py, and the gschema
files will be installed to $(DATADIR)\glib-2.0\schemas if DATADIR is
passed into the NMake command line, so that one can share this set
of gschema data files.  Note that the header file and introspection
files will still be installed to paths relative to $(PREFIX).