build: Use intermediate noinst library to generate gir
There's the peculiarity that we only want to generate the GIR for a header
file. The meson gnome module really wants an exec or a library. Since this
is a one off case, generate an intermediate noinst library so we keep meson
happy.

Still use the --header-only extra g-ir-scanner argument, so the GIR does
not ultimately depend on the library.