schema: Add "locate-pointer"
Initially, "locate pointer" was handled in gnome-settings-daemon using a
helper application which would work only on X11, not on Wayland.

While moving the feature to mutter/gnome-shell, however, the previous
setting in "org.gnome.settings-daemon.peripherals.mouse" is to be moved
to "org.gnome.desktop" instead.

However, having that setting in "org.gnome.desktop.peripherals.mouse" is
a bit too restrictive, because this is about the pointer and not just
the mouse (the device).

So add the new key "locate-pointer" in "org.gnome.desktop.interface" as
suggested by Carlos Garnacho.

https://gitlab.gnome.org/GNOME/gsettings-desktop-schemas/merge_requests/19
https://gitlab.gnome.org/GNOME/mutter/merge_requests/453
https://gitlab.gnome.org/GNOME/gnome-shell/issues/981
https://gitlab.gnome.org/GNOME/gnome-settings-daemon/merge_requests/86