schemas: add new lockdown setting for password showing
A convenient feature of many password entries is the
"Show Password" menu that lets the user see what password
is hidden behind the bullets they're typing.

While certainly useful, it can also be problematic in some
deployments, and violate security policies.

This commit adds a new setting to rid users of that "Show Password"
menu item.