build: Fix introspection on Visual Studio
Meson requires a library or an executable for building introspection files, but
unfortunately Visual Studio requires at least an object file to be built in
order to create a library or executable, which means a source file must be
present.

To satisfy this requirement, just use a dummy C source file on Visual Studio
builds and use it also to build our dummy library, so that we can generate the
introspection files properly, out of the box.