Fix description of use-same-proxy setting
Since 6cfbb00ac09e586a819181e7e39270569bb10a5a this setting is
supposedly unused, but in fact it has always been used. My archeology
has failed to turn up a reason for it to have been documented as unused.

I intend to remove all these settings eventually, because they are in
the wrong place, but in the meantime we might as well document them
properly.

Related: glib#2291