Fix default clock format, and allow localization
Currently the default clock format is unconditionally set to a 24-hour
clock. This is incorrect because the default locale is US English, which
uses a 12-hour clock. But we really do want to default to 24-hour clocks
in other locales. What to do?

Sebastian Keller proposed making the default value translatable, which
is apparently a thing that's actually possible. Well, sort of possible.
The functionality is currently broken, but it's *intended* to work, so
let's make use of it here. A fix for GLib to make this actually work
will be proposed separately, see glib#2575.

Translators: you will want to translate this from '12h' to '24h' if your
locale uses a 24-hour clock. Yes, it's a shame that we have to do this
when in theory the locale already knows how to format time, but this is
a pragmatic solution.

Fixes #30