a11y: Add new `show-status-shapes` setting
Whether or not switches include shapes to indicate their ON/OFF
state is currently controlled by the stylesheet (in particular
the HighContrast style).

However there are use cases for both using the HighContrast style
without shapes, and for using shapes with the regular stylesheet,
so in order to separate the two features, add a dedicated setting.

https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/2675