search: Include "Settings" in sort order
The special case that always sorted "Settings" first was dropped
in GNOME 44:

https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/1555
https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2584

The rationale for the change was to allow users to disable or
customize the provider, not to de-prioritize it and move it
towards the end of the list (providers not included in the
sort-order are sorted alphabetically).

The default should still be to show results from "Settings"
towards the top.