schemas: change defaults for stylus buttons to right and middle
The tip of a stylus works as left button virtually everywhere, binding
the physical buttons on the stylus to left button too makes the stylus
less useful out of the box. Map the first two stylus buttons to right
and middle, respectively.

The third stylus button is exceedingly rare anyway so let's leave that
one as-is.

Closes: https://gitlab.gnome.org/GNOME/gsettings-desktop-schemas/-/issues/71