schemas: Update default window button layout
Also, update the scheme's description.

The appmenu has been removed in GTK4 and before that removed from the
majority of all GNOME apps in GTK3 still. Appmenus never got a lot of
traction outside GNOME, one of the reasons why the were removed, so
it's pretty safe to say that theg are not commonly used anywhere
anymore. As a sidenote, Ubuntu has stopped defaulting to appmenus
around 2017.