schemas: Add eraser button configuration to the peripherals schemas
Many/most tablet tools have a built-in eraser button. Pressing this
button emulates the pen moving out of prox, being flipped and returned
to proximity with the eraser end. This is emulated at the HID level and
a requirement to pass MS certifications.

libinput 1.29 and later provide the ability to remap an eraser button to
behave like a normal button and send e.g. BTN_STYLUS. Let's provide the
configuration required for this as two-step: mapping the behavior and
then mapping the button if in button behavior.